+++ test/basics.test.js
@@ -528,6 +528,32 @@ var cases = [
         expect: { _args: [] }
     },
 
+    // env help
+    {
+        options: [
+            {names: ['a'], type: 'string', env: 'A', help: 'Phrase'},
+            {names: ['b'], type: 'string', env: 'B', help: 'Sentence.'},
+            {names: ['c'], type: 'string', env: 'C', help: 'Question?'},
+            {names: ['d'], type: 'string', env: 'D', help: 'Exclamation!'},
+            {names: ['e'], type: 'string', env: 'E', help: ' '},
+            {names: ['f'], type: 'string', env: 'F', help: ''},
+            {names: ['g'], type: 'string', env: 'G'},
+            {names: ['h'], type: 'bool', env: 'H'},
+        ],
+        argv: 'node tool.js --help',
+        helpOptions: { includeEnv: true },
+        expectHelp: [
+            /-a ARG\s+Phrase. Environment: A=ARG/,
+            /-b ARG\s+Sentence. Environment: B=ARG/,
+            /-c ARG\s+Question\? Environment: C=ARG/,
+            /-d ARG\s+Exclamation! Environment: D=ARG/,
+            /-e ARG\s+Environment: E=ARG/,
+            /-f ARG\s+Environment: F=ARG/,
+            /-g ARG\s+Environment: G=ARG/,
+            /-h\s+Environment: H=1/,
+        ]
+    },
+
     // env (number)
     {
         options: [ {names: ['timeout', 't'], env: 'FOO_TIMEOUT', type: 'number'} ],
@@ -607,11 +633,18 @@ var cases = [
 cases.forEach(function (c, i) {
     var expect = c.expect;
     delete c.expect;
-    var expectHelp = c.expectHelp;
-    if (typeof (expectHelp) === 'string') {
-        expectHelp = new RegExp(expectHelp);
+    var expectHelps = c.expectHelp;
+    if (!Array.isArray(expectHelps)) {
+        expectHelps = expectHelps ? [expectHelps] : [];
+        for (var i = 0; i < expectHelps.length; i++) {
+            if (typeof (expectHelps[i]) === 'string') {
+                expectHelps[i] = new RegExp(expectHelps[i]);
+            }
+        }
     }
     delete c.expectHelp;
+    var helpOptions = c.helpOptions;
+    delete c.helpOptions;
     var argv = c.argv;
     delete c.argv;
     if (typeof (argv) === 'string') {
@@ -653,10 +686,12 @@ cases.forEach(function (c, i) {
             debug('opts: %j', opts)
             t.deepEqual(opts, expect);
         }
-        if (expectHelp) {
-            var help = parser.help();
-            t.ok(expectHelp.test(help), format(
-                'help did not match %s: "%s"', expectHelp, help));
+        if (expectHelps.length) {
+            var help = parser.help(helpOptions);
+            expectHelps.forEach(function (eH) {
+                t.ok(eH.test(help), format(
+                    'help did not match %s: "%s"', eH, help));
+            });
         }
         t.end();
     });